afd: move behind feature

+++ src/scheduler/mod.rs
@@ -30,13 +30,19 @@ pub use drr::Drr;
 mod hdwrr;
 pub use hdwrr::HierarchicalDeficitWeightedRoundRobin;
 
+#[cfg(feature = "afd")]
 mod afd;
+#[cfg(feature = "afd")]
 pub use afd::ApproximateFairDropping;
 
+#[cfg(feature = "afd")]
 mod wafd;
+#[cfg(feature = "afd")]
 pub use wafd::WeightedApproximateFairDropping;
 
+#[cfg(feature = "afd")]
 mod hafd;
+#[cfg(feature = "afd")]
 pub use hafd::HierarchicalApproximateFairDropping;
 
 fn fnv(src: [u8; 4], dst: [u8; 4], queues: u64) -> u8 {
@@ -49,5 +55,5 @@ fn fnv(src: [u8; 4], dst: [u8; 4], queues: u64) -> u8 {
         hash = u64::wrapping_mul(hash, FNV_64_PRIME);
     }
 
-    (hash % queues as u64) as u8
+    (hash % queues) as u8
 }
